+++ e2e-tests/nightwatch.conf.js
@@ -16,7 +16,7 @@ https: nightwatch_config = {
   },
   // TODO: make it possible to pass an arg or have an environment variable or something
   // So that we actually test PRs and not just staging
-  launch_url: 'https://staging.ece-fawkes.ctoecskylight.com',
+  launch_url: 'https://qa.ece-fawkes.ctoecskylight.com',
   test_settings: {
     default: {
       globals: {
@@ -45,8 +45,7 @@ for (var i in nightwatch_config.test_settings) {
   var config = nightwatch_config.test_settings[i];
   config['selenium_host'] = nightwatch_config.selenium.host;
   config['selenium_port'] = nightwatch_config.selenium.port;
-  config.desiredCapabilities['browserstack.user'] =
-    process.env.BROWSERSTACK_USER;
+  config.desiredCapabilities['browserstack.user'] = process.env.BROWSERSTACK_USER;
   config.desiredCapabilities['browserstack.key'] = process.env.BROWSERSTACK_KEY;
 }
 module.exports = nightwatch_config;
