+++ e2e-tests/utils/acceptModal.js
@@ -1,9 +1,6 @@
 module.exports = {
   acceptModal: async function (browser, buttonText, noModalText) {
-    const modalButtonArgs = [
-      'xpath',
-      `//*/button[contains(., '${buttonText}')]`,
-    ];
+    const modalButtonArgs = ['xpath', `//*/button[contains(., '${buttonText}')]`];
     await browser.element(...modalButtonArgs, async (result) => {
       if (result.state === 'success') {
         await browser.click(...modalButtonArgs);
